+++ src/main/java/org/logicng/solvers/functions/ModelEnumerationFunction.java
@@ -56,7 +56,7 @@ import java.util.function.Consumer;
  * A solver function for enumerating models on the solver.
  * <p>
  * Model enumeration functions are instantiated via their builder {@link #builder()}.
- * @version 2.1.0
+ * @version 2.3.0
  * @since 2.0.0
  */
 public final class ModelEnumerationFunction implements SolverFunction<List<Assignment>> {
@@ -64,12 +64,14 @@ public final class ModelEnumerationFunction implements SolverFunction<List<Assig
     private final ModelEnumerationHandler handler;
     private final Collection<Variable> variables;
     private final Collection<Variable> additionalVariables;
+    private final boolean fastEvaluable;
 
     private ModelEnumerationFunction(final ModelEnumerationHandler handler, final Collection<Variable> variables,
-                                     final Collection<Variable> additionalVariables) {
+                                     final Collection<Variable> additionalVariables, final boolean fastEvaluable) {
         this.handler = handler;
         this.variables = variables;
         this.additionalVariables = additionalVariables;
+        this.fastEvaluable = fastEvaluable;
     }
 
     /**
@@ -127,7 +129,7 @@ public final class ModelEnumerationFunction implements SolverFunction<List<Assig
         }
         while (proceed && modelEnumerationSATCall(solver, this.handler)) {
             final LNGBooleanVector modelFromSolver = solver.underlyingSolver().model();
-            final Assignment model = solver.createAssignment(modelFromSolver, relevantAllIndices);
+            final Assignment model = solver.createAssignment(modelFromSolver, relevantAllIndices, this.fastEvaluable);
             models.add(model);
             proceed = this.handler == null || this.handler.foundModel(model);
             if (model.size() > 0) {
@@ -186,6 +188,7 @@ public final class ModelEnumerationFunction implements SolverFunction<List<Assig
         private ModelEnumerationHandler handler;
         private Collection<Variable> variables;
         private Collection<Variable> additionalVariables;
+        private boolean fastEvaluable = false;
 
         private Builder() {
             // Initialize only via factory
@@ -242,11 +245,21 @@ public final class ModelEnumerationFunction implements SolverFunction<List<Assig
         }
 
         /**
+         * Sets the flag whether the created assignment should be {@link Assignment#fastEvaluable() fast evaluable} assignments.
+         * @param fastEvaluable {@code true} if the created assignment should be fast evaluable, otherwise {@code false}
+         * @return the builder
+         */
+        public Builder fastEvaluable(final boolean fastEvaluable) {
+            this.fastEvaluable = fastEvaluable;
+            return this;
+        }
+
+        /**
          * Builds the model enumeration function with the current builder's configuration.
          * @return the model enumeration function
          */
         public ModelEnumerationFunction build() {
-            return new ModelEnumerationFunction(this.handler, this.variables, this.additionalVariables);
+            return new ModelEnumerationFunction(this.handler, this.variables, this.additionalVariables, this.fastEvaluable);
         }
     }
 }